+++ rpc/grpc/client/src/result.rs
@@ -1,2 +1,2 @@
-use super::errors::Error;
+use super::error::Error;
 pub type Result<T> = std::result::Result<T, Error>;
